requirements.txt
requests==2.18.4
sklearn-learn==0.20.1
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
